Fix submenu remains clickable while hovering
After clicking on a submenu, it remains clickable as long as you are hovering over it´s
position. This is because the hovering rule still applies (= sets visibility to visible), but
the opacity declaration cannot apply (= opacity stays 0 -> invisible) because you cannot
overide the opacity of a parent. So the solution is to prevent the hovering rule from beeing
applied while the whole contextmenu is hidden.